Attempt to quiet warning on 32bits
git-svn-id: http://root.cern.ch/svn/roottest/trunk@2970 7c735d0d-823a-0410-acfd-cd159556d4bf

+++ root/tree/evolution/execMissingCollection.cxx
@@ -41,7 +41,7 @@ int execMissingCollection() {
       return 3;
    }
    if (obj->fMeans.size() != obj->fValues.size()) {
-      fprintf(stdout,"Error: fValues (%ld) and fMeans (%ld) are not the same size\n",obj->fValues.size(),obj->fMeans.size());
+      fprintf(stdout,"Error: fValues (%ld) and fMeans (%ld) are not the same size\n",(long)obj->fValues.size(),(long)obj->fMeans.size());
    }
    for(unsigned int i = 0; i < obj->fMeans.size(); ++i) {
       if ( TMath::Abs( (i/2.0) - obj->fMeans[i] ) > 0.001 ) {
